refactor: extract shared getInitials and getBaseUrl helpers
- lib/user.ts: getInitials() extracted from workspace and admin layouts
- server/request.ts: getBaseUrl() extracted from admin/invites and
  admin/users (both had identical x-forwarded-proto/host derivation inline)
- api/admin/jobs GET: add enforceSameOrigin to match pattern set by
  the POST /run route
