+++ apps/web/app/admin/layout.tsx
@@ -1,20 +1,10 @@
+import { getInitials } from "@/lib/user";
 import { requireOwnerPageSession } from "@/server/auth/guards";
 
 import { AdminNav } from "./admin-nav";
 import { AdminTopbarActions } from "./admin-topbar-actions";
 
 export const dynamic = "force-dynamic";
-
-function getInitials(displayName: string | null, username: string): string {
-  if (displayName) {
-    const parts = displayName.trim().split(/\s+/);
-    if (parts.length >= 2) {
-      return (parts[0]![0]! + parts[1]![0]!).toUpperCase();
-    }
-    return displayName.slice(0, 2).toUpperCase();
-  }
-  return username.slice(0, 2).toUpperCase();
-}
 
 export default async function AdminLayout({
   children,
