+++ apps/web/app/api/admin/jobs/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 
-import { requireOwnerApiSession } from "@/server/admin/http";
+import { enforceSameOrigin, requireOwnerApiSession } from "@/server/admin/http";
 import {
   getAdminJobList,
   parseAdminJobFilters,
@@ -8,6 +8,9 @@ import {
 } from "@/server/admin/jobs";
 
 export async function GET(request: NextRequest) {
+  const sameOriginError = enforceSameOrigin(request);
+  if (sameOriginError) return sameOriginError;
+
   const auth = await requireOwnerApiSession(request);
 
   if (!auth.ok) {
